% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdd.R
\name{get_mdd}
\alias{get_mdd}
\title{Function to return mammal shapefiles of Rodentia order}
\source{
Expert range maps of global mammal distributions harmonised to
three taxonomic authorities
\url{https://doi.org/10.1111/jbi.14330}
}
\usage{
get_mdd(dir = NULL)
}
\arguments{
\item{dir}{A directory where to write the output}

\item{Order}{The mammal order want to  download}
}
\value{
A SpatVector object with mammal shapefiles
}
\description{
Function to return mammal shapefiles of Rodentia order
}
\examples{
get_mdd()
}
