% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdd_chiroptera.R
\name{get_mdd_chiroptera}
\alias{get_mdd_chiroptera}
\title{Function to return mammal shapefiles of Chiroptera Order}
\source{
Expert range maps of global mammal distributions harmonised to
 three taxonomic authorities
 \url{https://doi.org/10.1111/jbi.14330}
}
\usage{
get_mdd_chiroptera(dir = NULL)
}
\arguments{
\item{dir}{A directory where to write the output}
}
\value{
A SpatVector object with mammal shapefiles
}
\description{
Function to return mammal shapefiles of Chiroptera Order
}
\examples{
get_mdd_chiroptera()
}
