% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdd_map.R
\name{get_mdd_map}
\alias{get_mdd_map}
\title{Title}
\source{
Expert range maps of global mammal distributions harmonised to
three taxonomic authorities
\url{https://doi.org/10.1111/jbi.14330}
}
\usage{
get_mdd_map(
  species = NULL,
  Order = NULL,
  country = NULL,
  continent = NULL,
  cropCountry = FALSE
)
}
\arguments{
\item{species}{A vector of mammal species name.}

\item{Order}{A vector of mammal orders to retrive.}

\item{country}{A string of country name in English.}

\item{continent}{A string of continent name in English.}

\item{cropCountry}{Logical. If TRUE, the shapefiles are crop with
country or continent shape}
}
\value{
A SpatVector object with mammal shapefiles
}
\description{
Title
}
\examples{
get_mdd_map("Notoryctes caurinus")
}
