\name{MDD-internal}
\title{Internal MDD objects}
\alias{.all.less}
\alias{.all.subs}
\alias{.compress.list}
\alias{.explanation.mantelhaen}
\alias{.mantelhaen}
\alias{.maxa}
\alias{.meld.strata}
\alias{.mhstat}
\alias{.mina}
\alias{.one.less}
\alias{.one.more}
\alias{.validate.entries}
\description{Internal MDD objects.}
\details{These are not to be called by the user.}
\keyword{internal}