\name{MDD-package}
\alias{MDD-package}
\alias{MDD}
\docType{package}
\title{
Minimum Detectable Difference Calculations}
\description{
Tools for calculating the minmum detectable difference (MDD) for various tests with 
continuous or binary outcomes.

Also contains programs to compare the MDD to the clinically significant difference 
for most of the same tests.
}
\details{
\tabular{ll}{
Package: \tab MDD\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2008-06-20\cr
License: \tab GPL-2\cr
}
}
\author{
Don Barkauskas

Maintainer: Don Barkauskas (\email{barkda@wald.ucdavis.edu})
}
\keyword{ package }
