% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mddSpListSample.R
\docType{data}
\name{mddSpListSample}
\alias{mddSpListSample}
\title{Species list sample from Mammal Diversity Database}
\format{
\subsection{\code{mddSpListSample}}{

A data frame with 5 rows and 8 columns:
\describe{
\item{Order}{Order of the species}
\item{Family}{Family of the species }
\item{Genus}{Genus of the species}
\item{Species}{Species Name}
\item{MDD_SciName}{Species binomial name}
\item{Digitisable}{Is digitisable?}
\item{diffSinceCMW}{Has a difference from CMW}
\item{CMW_SciName}{Species name in CMW }
\item{shp_orig}{Source of the shp}
...
}
}
}
\source{
Expert range maps of global mammal distributions harmonised to
three taxonomic authorities
\url{https://doi.org/10.1111/jbi.14330}
}
\usage{
mddSpListSample
}
\description{
A list of species related with shapefiles
in MDD. Includes taxonomy
}
\keyword{datasets}
