% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{url_exists}
\alias{url_exists}
\title{Url Exists}
\usage{
url_exists(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{A string with the url to check if exist}

\item{quiet}{A logic if quiets}

\item{...}{Ellipsis}
}
\value{
A logic value if url exists
}
\description{
Url Exists
}
\keyword{internal}
