% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdd_Rodentia.R
\name{get_mdd_Rodentia}
\alias{get_mdd_Rodentia}
\title{Function to return mammal shapefiles of afrosoricida Order}
\source{
A phylogeny-informed characterization of global tetrapod traits
 addresses data gaps and biases
 \url{https://doi.org/10.1101/2023.03.04.531098}
}
\usage{
get_mdd_Rodentia(dir = NULL)
}
\arguments{
\item{dir}{A directory where to write the output}
}
\value{
A SpatVector object with mammal shapefiles
}
\description{
Function to return mammal shapefiles of afrosoricida Order
}
\examples{
get_mdd_afrosoricida()
}
